generalize fixtures for cli tests
